Update Imap.php
- Fixed 2 spelling mistakes
- Potential fix for a weird IMAP Authentication bug

When using IMAP Authenticator module, sometimes an invalid password errors out in 2 different ways. I believe there may be a cache involved somewhere but I'm not sure.

The first error is reported fine to the end-user in the GUI and it gets generated in 
groupoffice/api/auth.php line 118
The error gets sent back as an HTTP status code: HTTP/1.1 401 Bad username or password

This works fine

Now in the second case, it performs the authentication all the way to the IMAP server which goes to this file. The error gets generated like so:

"Authententication failed for user user@domain on IMAP server imap.domain.com

A3 NO [ AUTHENTICATIONFAILED ] Authentication failed"

Now that error gets sent back all the way back to auth.php which attempts to call setStatus from core/http/Response.php
In that file/function we have the line 'header($status)' which fails with this error: Header may not contain more than a single header, new line detected

Removing those newlines + the IMAP response fixes this problem.

I'm not sure why there are 2 different paths for IMAP authentication, in any case, for a password failure sometimes it goes through path #1 , sometimes through this path #2. The errors are somewhat different too.

For path #1 it shows status code 401 , message: Bad username or password
For path #2 it shows status code 500, Sorry, an error occurred. Authentication failed for user...

This fix takes care of the issue in #2 but it would be nice if both password failures could be handled similarly