Update HtmlEditor.js
Major rewrite of the GO HtmlEditor component. Please note that our original code was based on version 6.7.78 which doesn't include this commit:

https://github.com/Intermesh/groupoffice/commit/9c3ef5c50c34ae526bdd0ea105ba12067a1a34cc

The main issue we tried solving was to have the most common buttons more readily available and visible, to have a proper functioning overflow menu in all window sizes and devices. The image button for example was added as a plugin but way too far to the right and not even visible in the Template Editor for example.

We fixed all this by:

- Removed some plugins like the wordpaste plugin, ssplugin, rmFormat. These in our experience were very rarely used if at all. Pasting from Word works great even without that plugin.

- combining the B/I/U buttons as one and allowing them to be checked in order to apply the format and also to be updated depending on the cursor position.
- moved the insert image button right after it
- insert table/emojis are next
- added realtime indicator of the font color.
- removed all icons except font family/size B/I/U and insert image/emoijs from mobile devices. All other functions are spotty at best on mobile and I haven't seen other webmail to offer them. The main problem is returning the focus and inserting at the right place the  various items/action.

Group-Office overrides the ExtJS toolbar template with two spans, one for the icon, one for the text of the icon. But this doesn't play nice at all with the overflow menu which would show [Object object] since it was trying to show both icon and text.

Lots of the code is just handling such small quirks.